+++ dotnet/Data/IPackingRepository.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using FedexShipping.Models;
+
+namespace FedexShipping.Data
+{
+    public interface IPackingRepository
+    {
+        Task<PackingResponseWrapper> PackItems(PackingRequest packingRequest);
+    }
+}