+++ dotnet/Services/IPackingService.cs
@@ -1,0 +1,11 @@
+﻿using FedexShipping.Models;
+using System.Threading.Tasks;
+using System.Collections.Generic;
+
+namespace FedexShipping.Services
+{
+    public interface IPackingService
+    {
+        Task<List<Item>> packingMap(List<Item> items);
+    }
+}